move Camera to extra file
